Add WTKD support to Image.LoadTask.
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@770011 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/pivot/wtk/media/drawing/Ellipse.java
@@ -38,7 +38,7 @@ public class Ellipse extends Shape {
         }
     }
 
-    private Ellipse2D.Double ellipse2D = new Ellipse2D.Double();
+    private Ellipse2D.Float ellipse2D = new Ellipse2D.Float();
 
     private EllipseListenerList ellipseListeners = new EllipseListenerList();
